[optimize] 代码结构优化
